Moved chan sender/receiver to protocol/test
